refactor(combobox): improved clarity of useCombobox hook
